feat(hotels): get available vacancies
